// maravich/src/Utils/types.ts
export interface Player {
  firstName: string
  lastName: string
  playerId: number
}

export interface Stats {
  assists: number
  blocks: number
  defReb: number
  fga: number
  fgm: number
  fgp: number
  fta: number
  ftm: number
  ftp: number
  min: number
  offReb: number
  pFouls: number
  playerId: number
  plusMinus: number
  points: number
  steals: number
  teamId: number
  totReb: number
  tpa: number
  tpm: number
  tpp: number
  turnovers: number
}